+++ typings.d.ts
@@ -2,9 +2,3 @@ declare module '*.json' {
   const value: any
   export default value
 }
-
-declare module '*.vue' {
-  import { Component } from 'vue'
-  const Comp: Component
-  export default Comp
-}